% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{generate_example}
\alias{generate_example}
\title{'generate_example': Generate the data examples presented in Goldberg and Gorfine (2019)}
\usage{
generate_example(n = 100, beta = c(-1, -1), setting = 1,
  parameters = 1, seed = -1)
}
\arguments{
\item{n}{Sample size}

\item{beta}{A 2x1 vector of coefficeients.}

\item{setting}{The setting number is a number between 1 and 4.
See the Simulation Section of Goldberg and Gorfine (2019) for the description of the settings.}

\item{parameters}{The parameters of the example. The parameters depend on the example's setting.
For setting 1,3, and 4 the parameter that was used is 0.25.
For setting 2, the parameter that was used is 3}

\item{seed}{Use the seed in creating the example. If seed is negetive, no seed is used.}
}
\value{
A list that contains the nx2 matrix of explanatory variables x, the response vector y.
If applicable also the missingness indicator R and a vector of weights.
}
\description{
'generate_example': Generate the data examples presented in Goldberg and Gorfine (2019)
}
