% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr.R
\name{lpr}
\alias{lpr}
\title{lpr: Main function for calculating the model parameters for a proportional likelihood model}
\usage{
lpr(x, y, method, weights = NULL)
}
\arguments{
\item{x}{An nxp matrix of the explanatory variables}

\item{y}{An nx1 vector of response}

\item{method}{The method to use when calculating the parameters:
"1" or "Lou.Tsai" for the method of Lou & Tsai (2012),
"2" or "Chan" for the method of Chen (2013),
"3" or "Gold.Gorf" for the method of Goldberg and Gorfine (2019).
Method of Gold.Gorf can handle survival. Methods Chan and Gold.Gorf can handle missing data.}

\item{weights}{Vector of weights for missing data or survival. Can be used only with Gold.Gorf method.
The weights should be the inverse of the probability of missingness or censoring.}
}
\value{
A list that contains a px1 vector beta and the baseline distribution G (see Lou & Tsai, 2012, for details)
}
\description{
lpr: Main function for calculating the model parameters for a proportional likelihood model
}
