% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{message_to_test_result}
\alias{message_to_test_result}
\title{Helper function to format result object returned to PL}
\usage{
message_to_test_result(msg, max_pts = 100)
}
\arguments{
\item{msg}{Character variable with the error or warning received}

\item{max_pts}{Optional numeric variable with maximal attainable points, usual 100}
}
\value{
A data.frame object with four elements as expected by PL
}
\description{
If there is an error, format the error message so that it can be displayed in
the PL question details.
}
