% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source_as}
\alias{source_as}
\title{Source a file as a particular user}
\usage{
source_as(file, uid = "ag")
}
\arguments{
\item{file}{Path to an R file to be sourced, typically containing student
code to be evaluated safely, that is, with lower privileges.}

\item{uid}{Optional numeric or character user id identifying the user id with
(presumably lower) privileges as which the code is running. The corresponding
numeric \code{uid} is obtained via \code{\link[unix]{user_info}} when a
character \code{uid} is supplied. Note that using this argument requires
being the \sQuote{root} user. Default value of \code{"ag"} will source
supplied code with privileges that prevent access to sensitive files such as
reference solutions.}
}
\value{
An environment containing objects created through sourcing the
\code{file} as \code{uid}. Any sourced functions are modified such that
future use of the function will be restricted by the lower privileges given
to the supplied \code{uid}.
}
\description{
Source a file as a particular user
}
\examples{
\dontrun{
std <- source_as(file = "/grade/student/sub.R")
ref <- source_as(file = "/grade/student/ans.R")
}
}
