% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source_as_uid}
\alias{source_as_uid}
\alias{source_std}
\alias{source_ref}
\title{Source a file as a particular user}
\usage{
source_as_uid(file, uid = "ag")

source_std(name, uid = "ag")

source_ref(name, uid = "ag")
}
\arguments{
\item{file}{Path to an R file to be sourced, typically containing student
code to be evaluated safely, that is, with lower privileges.}

\item{uid}{Optional numeric or character user id identifying the user id with
(presumably lower) privileges as which the code is running. The corresponding
numeric \code{uid} is obtained via \code{\link[unix]{user_info}} when a
character \code{uid} is supplied. Note that using this argument requires
being the \sQuote{root} user. Default value of \code{"ag"} will source
supplied code with privileges that prevent access to sensitive files such as
reference solutions.}

\item{name}{Name of file (usually student or reference solution) to be
sourced.}
}
\value{
An environment containing objects created through sourcing the
\code{file} as \code{uid}. Any sourced functions are modified such that
future use of the function will be restricted by the lower privileges given
to the supplied \code{uid}.
}
\description{
The \code{source_std} helper sources a user-specified file from the
\code{/grade/student/} directory as the user \code{"ag"}. Similarly, the
\code{source_std} helper sources a user-specified file from the
\code{/grade/tests/} directory as the user \code{"ag"}
}
\examples{
\dontrun{
std <- source_as_uid(file = "/grade/student/sub.R", uid = "ag")
ref <- source_as_uid(file = "/grade/student/ans.R", uid = "ag")
std <- source_std("sub.R")
ref <- source_ref("ans.R")
}
}
