% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{test_question}
\alias{test_question}
\title{Run a whole question and report aggregate results}
\usage{
test_question(tests_dir = "/grade/tests/tests", results_file = "results.json")
}
\arguments{
\item{tests_dir}{Directory containing the test files for a question}

\item{results_file}{JSON file into which results are written}
}
\value{
The results data.frame is returned, but the functions is invoked for its
side-effect of creating the JSON file
}
\description{
This function is the equivalent of the \code{pltest.R} script which, given a directory
runs the tests file therein and reports the results in a JSON file for PrairieLearn to
consume.
}
